requirements.txt
dnspython==1.14.0
Flask==0.11.1